chore: using signed_zero_F32, signed_inf_F32. Proper true/false operands for bool instead 0/1

+++ source/others/s_normRoundPackToF32.c
@@ -41,7 +41,7 @@ softfloat_normRoundPackToF32(bool sign, int16_t exp, uint32_t sig)
 {
     int8_t const shiftDist = softfloat_countLeadingZeros32(sig) - 1;
     exp -= shiftDist;
-    if ((7 <= shiftDist) && ((unsigned int)exp < 0xFD)) {
+    if (7 <= shiftDist && (uint16_t)exp < 0xFD) {
         return u_as_f_32(packToF32UI(sign, sig ? exp : 0, sig << (shiftDist - 7)));
     } else {
         return softfloat_roundPackToF32(sign, exp, sig << shiftDist);
